azure_functions/requirements.txt
azure-functions==1.17.0
pyodbc==4.0.39